move `Common::String` as `Shell::InputWindow::EventfulBuffer`
