+++ sandman/source/shell/attributes.h
